# travel-journal/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Travel Experience Journal API",
    "description": "REST API za platformu za deljenje putničkih iskustava",
    "version": "1.0.0",
    "contact": {
      "name": "Travel Journal Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Lokalni development server"
    }
  ],
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registracija novog korisnika",
        "description": "Kreira novog korisnika u Firebase Auth i Firestore",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Korisnik uspešno registrovan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validaciona greška ili korisnik već postoji",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server greška",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ratings": {
      "get": {
        "tags": [
          "Ratings"
        ],
        "summary": "Ocene za destinaciju",
        "parameters": [
          {
            "name": "destinationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ocene za destinaciju",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ratings"
        ],
        "summary": "Oceni destinaciju",
        "description": "Korisnik može dati jednu ocenu po destinaciji (1-5). Ponovni poziv ažurira postojeću ocenu.",
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRatingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ocena uspešno sačuvana"
          },
          "400": {
            "description": "Validaciona greška"
          },
          "401": {
            "description": "Neautentifikovan"
          }
        }
      }
    },
...
    "/api/profile": {
      "put": {
        "tags": [
          "Profile"
        ],
        "summary": "Ažuriranje profila",
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil uspešno ažuriran"
          },
          "400": {
            "description": "Validaciona greška"
          },
          "401": {
            "description": "Neautentifikovan"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth JWT session cookie"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Greška pri obradi zahteva"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "password",
          "displayName"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "korisnik@primer.com"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "example": "mojasifra123"
          },
          "displayName": {
            "type": "string",
            "minLength": 2,
            "example": "Marko Marković"
          }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "user": {
            "type": "object",
            "properties": {
              "uid": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "displayName": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "RatingsResponse": {
        "type": "object",
        "properties": {
          "ratings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rating"
            }
          },
          "averageRating": {
            "type": "number",
            "example": 4.3
          },
          "totalRatings": {
            "type": "integer",
            "example": 7
          },
          "userRating": {
            "type": "integer",
            "nullable": true,
            "example": 5
          }
        }
      },
      "Rating": {
        "type": "object",
        "properties": {
          "ratingId": {
            "type": "string"
          },
          "destinationId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "score": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateRatingRequest": {
        "type": "object",
        "required": [
          "destinationId",
          "score"
        ],
        "properties": {
          "destinationId": {
            "type": "string",
            "example": "dest123"
          },
          "score": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 4
          }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "required": [
          "displayName"
        ],
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 2,
            "example": "Marko Marković"
          },
          "bio": {
            "type": "string",
            "example": "Strastveni putnik"
          },
          "profilePhotoURL": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/photo.jpg"
          }
        }
      }
    }
  }
}